ダッシュボード画面からActivityChartコンポーネントを削除し、WeeklyActivityChartViewに置き換えました。また、ActivityChart内の不要なコメントやコードを整理しました。これにより、コードの可読性が向上し、ダッシュボードの表示がシンプルになりました。
