+++ feature/dashboard/presentation/components/display/charts/ActivityChart.tsx
@@ -1,6 +1,5 @@
 import React, { useEffect } from "react";
 import { Animated, View } from "react-native";
-import { TrendingUpIcon } from "lucide-react-native";
 
 import { ThemedText } from "@/components/text/ThemedText";
 
@@ -93,24 +92,6 @@ export const ActivityChart = ({ data }: ActivityChartProps) => {
 
   return (
     <View className="w-full rounded-xl dark:bg-gray-900">
-      {/* Title section */}
-      <View className="mb-6 flex-row items-center justify-between">
-        <View className="flex-row items-center gap-2">
-          <TrendingUpIcon size={16} color="#FA4714" strokeWidth={1.5} />
-          <ThemedText
-            text={"Weekly Activity"}
-            variant="body"
-            weight="semibold"
-            color="default"
-          />
-        </View>
-        <ThemedText
-          text={"2025/03/16~2025/03/22"}
-          variant="caption"
-          weight="normal"
-          color="muted"
-        />
-      </View>
       {/* Chart section - custom implementation */}
       <View
         style={{ height: chartHeight }}
@@ -204,184 +185,6 @@ export const ActivityChart = ({ data }: ActivityChartProps) => {
             </View>
           );
         })}
-      </View>
-
-      {/* 凡例 - iOSスタイル */}
-      <View className="mt-2 flex-row justify-end gap-4 px-2">
-        <View className="flex-row items-center">
-          <View
-            className="mr-2 h-3 w-3 rounded-sm"
-            style={{ backgroundColor: colors.add.main }}
-          />
-          <ThemedText
-            text={`add`}
-            variant="small"
-            color="muted"
-            className="text-xs"
-          />
-        </View>
-
-        <View className="flex-row items-center">
-          <View
-            className="mr-2 h-3 w-3 rounded-sm"
-            style={{ backgroundColor: colors.swipe.main }}
-          />
-          <ThemedText
-            text={`swipe`}
-            variant="small"
-            color="muted"
-            className="text-xs"
-          />
-        </View>
-
-        <View className="flex-row items-center">
-          <View
-            className="mr-2 h-3 w-3 rounded-sm"
-            style={{ backgroundColor: colors.read.main }}
-          />
-          <ThemedText
-            text={`read`}
-            variant="small"
-            color="muted"
-            className="text-xs"
-          />
-        </View>
-      </View>
-
-      {/* カテゴリ別合計時間 - 表形式 */}
-      <View className="mt-2 border-t border-zinc-300 px-2 pt-4 dark:border-zinc-700">
-        {/* ヘッダー行 */}
-        <View className="mb-4 flex-row">
-          <View className="flex-1">
-            <ThemedText
-              text="Actions"
-              variant="small"
-              weight="semibold"
-              color="muted"
-              className="text-xs"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text="Average"
-              variant="small"
-              weight="semibold"
-              color="muted"
-              className="text-right text-xs"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text="Total"
-              variant="small"
-              weight="semibold"
-              color="muted"
-              className="text-right text-xs"
-            />
-          </View>
-        </View>
-
-        {/* add行 */}
-        <View className="mb-4 flex-row items-center">
-          <View className="flex-1 flex-row items-center">
-            <View
-              className="mr-2 h-3 w-3 rounded-sm"
-              style={{ backgroundColor: colors.add.main }}
-            />
-            <ThemedText
-              text="add"
-              variant="body"
-              weight="medium"
-              color="default"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`15`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`30`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-        </View>
-
-        {/* swipe行 */}
-        <View className="mb-4 flex-row items-center">
-          <View className="flex-1 flex-row items-center">
-            <View
-              className="mr-2 h-3 w-3 rounded-sm"
-              style={{ backgroundColor: colors.swipe.main }}
-            />
-            <ThemedText
-              text="swipe"
-              variant="body"
-              weight="medium"
-              color="default"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`12`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`30`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-        </View>
-
-        {/* read行 */}
-        <View className="flex-row items-center">
-          <View className="flex-1 flex-row items-center">
-            <View
-              className="mr-2 h-3 w-3 rounded-sm"
-              style={{ backgroundColor: colors.read.main }}
-            />
-            <ThemedText
-              text="read"
-              variant="body"
-              weight="medium"
-              color="default"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`18`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-          <View className="w-20">
-            <ThemedText
-              text={`30`}
-              variant="body"
-              weight="medium"
-              color="muted"
-              className="text-right"
-            />
-          </View>
-        </View>
       </View>
     </View>
   );
